Adds support for bit operations in the program track API (#44)
Adds an encoding for the program track CV space representation that allows expressing a single-bit write.
Implements this in the programming track CV space. Adds API for POM bit operations to the program track frontend, but this is not implemented.
Currently bit operations will only work on the program track. Reads are not supported, just writes.

* Adds API for writing bits via POM. (no implementation though)

* Adds support for bit operations via the high address bits in the DCC CV space.