Optimize the amount of time we spend in repeated searches. (#54)
Adds a heuristic for detecting when a throttle sends multiple searches one after the other. In this case the previous search will be canceled and we move on to the processing of the new search directly.

Adds some debugging printout for latency understanding.

====

* Optimize the amount of time we spend in repeated searches.

Adds a heuristic for detecting when a throttle sends multiple searches
one after the other. In this case the previous search will be canceled
and we move on to the processing of the new search directly.

Adds some debugging printout for latency understanding.

* Fix whitespace.

* Adds cleaner documentation comments.