Adds POM support for accessory decoders (#61)
Adds a new mode "POM with address" to the CV space. (Mode = top 8 bits of the memconfig address.)

When using POM with address, there will be 10-bit CV number, 14-bit address, and 8-bit mode selection in the memconfig address. The 14-bit address can be used to target a locomotive or an accessory decoder. This will all work when talking to the CS's node id.

===

* Adds support for ACCY POM in the CV space and respective frontend.

* Changes how POM selection is handled. Instead of a separate mode for each
address type (e.g. pom for mobile decoder, POM for basic accessory, POM for
extended accessory) there is now a generic POM mode, and then an address
type selection allows determining what kind of packet to generate.

This rearranges the entire high side of the programming track space.

Now we can send POM requests on the CS's node id, if those POM requests are
addressed to an accessory, they will work.

* Enables POM with address even when service mode is disabled.