Increase the number of program track startup reset packets (#62)
Some decoders that have a lot of onboard capacitance, and are very close to the program track steady state current limits, can experience an increase in current following an ack. This can result sometimes result in a current fault being detected.

Increasing the number of reset packets when the program track starts provides a bit more time for the onboard capacitance to charge up ahead of the first ack. This provides a bit of extra discharge capacity when the ack comes, preventing the expected uptick in charging current following the ack from triggering a current fault.

===

* Increase the number of reset packers on at the startup of the program track.

* Update comment with explanation.